removed warning and updated Ram usage for process to be int

+++ CppND-System-Monitor/src/process.cpp
@@ -42,7 +42,7 @@ string Process::Ram()
     // Convert to MB
     try
     {
-        ram = std::to_string(stof(ram) / 1024);
+        ram = std::to_string(stol(ram) / 1024);
     }
     catch(const std::exception& e)
     {
